+++ docs/guide/developer.rst
@@ -19,7 +19,9 @@ The library is not meant to be modular, although inheritance is used to reduce c
 Algorithms Structure
 ====================
 
+
 Each algorithm (on-policy and off-policy ones) follows a common structure.
+Policy contains code for acting in the environment, and algorithm updates this policy.
 There is one folder per algorithm, and in that folder there is the algorithm and the policy definition (``policies.py``).
 
 Each algorithm has two main methods:
@@ -34,13 +36,14 @@ Where to start?
 
 The first thing you need to read and understand are the base classes in the ``common/`` folder:
 
-- ``BaseRLModel`` in ``base_class.py`` which defines how an RL class should look like.
+- ``BaseAlgorithm`` in ``base_class.py`` which defines how an RL class should look like.
   It contains also all the "glue code" for saving/loading and the common operations (wrapping environments)
 
 - ``BasePolicy`` in ``policies.py`` which defines how a policy class should look like.
-  It contains also all the magic for the ``.predict()`` method, to handle as many cases as possible
+  It contains also all the magic for the ``.predict()`` method, to handle as many spaces/cases as possible
 
-- ``OffPolicyRLModel`` in ``base_class.py`` that contains the implementation of ``collect_rollouts()`` for the off-policy algorithms
+- ``OffPolicyAlgorithm`` in ``off_policy_algorithm.py`` that contains the implementation of ``collect_rollouts()`` for the off-policy algorithms,
+  and similarly ``OnPolicyAlgorithm`` in ``on_policy_algorithm.py``.
 
 
 All the environments handled internally are assumed to be ``VecEnv`` (``gym.Env`` are automatically wrapped).
@@ -50,7 +53,7 @@ Pre-Processing
 ==============
 
 To handle different observation spaces, some pre-processing needs to be done (e.g. one-hot encoding for discrete observation).
-Most of the code for pre-processing is in ``common/preprocessing.py``.
+Most of the code for pre-processing is in ``common/preprocessing.py`` and ``common/policies.py``.
 
 For images, we make use of an additional wrapper ``VecTransposeImage`` because PyTorch uses the "channel-first" convention.
 
@@ -61,8 +64,11 @@ Policy Structure
 When we refer to "policy" in Stable-Baselines3, this is usually an abuse of language compared to RL terminology.
 In SB3, "policy" refers to the class that handles all the networks useful for training,
 so not only the network used to predict actions (the "learned controller").
-
 For instance, the ``TD3`` policy contains the actor, the critic and the target networks.
+
+To avoid the hassle of importing specific policy classes for specific algorithm (e.g. both A2C and PPO use ``ActorCriticPolicy``),
+SB3 uses names like "MlpPolicy" and "CnnPolicy" to refer policies using small feed-forward networks or convolutional networks,
+respectively. Importing ``[algorithm]/policies.py`` registers an appropriate policy for that algorithm under those names.
 
 Probability distributions
 =========================
