+++ docs/modules/base.rst
@@ -8,14 +8,29 @@ Base RL Class
 
 Common interface for all the RL algorithms
 
-.. autoclass:: BaseRLModel
+.. autoclass:: BaseAlgorithm
   :members:
+
+
+.. automodule:: stable_baselines3.common.off_policy_algorithm
 
 
 Base Off-Policy Class
 ^^^^^^^^^^^^^^^^^^^^^
 
-The base RL model for Off-Policy algorithm (ex: SAC/TD3)
+The base RL algorithm for Off-Policy algorithm (ex: SAC/TD3)
 
-.. autoclass:: OffPolicyRLModel
+.. autoclass:: OffPolicyAlgorithm
+  :members:
+
+
+.. automodule:: stable_baselines3.common.on_policy_algorithm
+
+
+Base On-Policy Class
+^^^^^^^^^^^^^^^^^^^^^
+
+The base RL algorithm for On-Policy algorithm (ex: A2C/PPO)
+
+.. autoclass:: OnPolicyAlgorithm
   :members:
