+++ stable_baselines3/a2c/policies.py
@@ -1,5 +1,5 @@
 # This file is here just to define MlpPolicy/CnnPolicy
-# that work for PPO
+# that work for A2C
 from stable_baselines3.common.policies import ActorCriticPolicy, ActorCriticCnnPolicy, register_policy
 
 MlpPolicy = ActorCriticPolicy
